requirements.txt
tensorflow-plot
tensorflow-ops
mesh-tensorflow
tenfsorflow-io
tensorflow-recommenders
tensorflow-scientific
tensorflow-tracer
tensorflow-tracer
tensorflow-manopt
tensorflow-probability
tensorflow-model
tensorflow-gan
tensorflow-determinisim
tensorflow-utils
tensorflow-ranking
tensorflow-cpu-estimator
tensorflow-lattice
tensorflow-graph
tensorflow-rl
tensorflow-quantum
tensorflow-encrypted
tensorflow-text
tensorflow-model-remediation
tensorflow-model-analysis
tensorflow-transform-canary
tensorflow-hub
